Side logic

+++ web/backend/communect/src/main/kotlin/com/example/communect/ui/controller/TalkAPIController.kt
@@ -1,7 +1,7 @@
 package com.example.communect.ui.controller
 
-import com.example.communect.app.service.MockTestData
 import com.example.communect.domain.model.IndividualTalkIns
+import com.example.communect.domain.model.Login
 import com.example.communect.domain.model.MessageIns
 import com.example.communect.domain.model.TalkUpd
 import com.example.communect.domain.service.MessageService
@@ -9,6 +9,7 @@ import com.example.communect.domain.service.TalkService
 import com.example.communect.ui.form.*
 import org.apache.coyote.BadRequestException
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.validation.BindingResult
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.*
@@ -22,26 +23,31 @@ class TalkAPIController(
 ) {
     /** 個人トーク一覧取得 */
     @GetMapping
-    fun getIndividualTalks(): TalksResponse {
-        val talks = talkService.getIndividualTalks(MockTestData.user1.userId)
+    fun getIndividualTalks(
+        @AuthenticationPrincipal loginUser: Login
+    ): TalksResponse {
+        val talks = talkService.getIndividualTalks(loginUser.user.userId)
         return TalksResponse(talks.map { TalkInfo(it) })
     }
 
-    /** 個人トーク一覧取得 */
+    /** トーク取得 */
     @GetMapping("/{talkId}")
     fun getIndividualTalk(
+        @AuthenticationPrincipal loginUser: Login,
         @PathVariable("talkId") talkId: String
     ): TalkResponse {
-        val talk = talkService.getTalk(talkId) ?: throw BadRequestException()
+        val talk = talkService.getTalk(talkId, loginUser.user.userId) ?: throw BadRequestException()
         return TalkResponse(TalkInfo(talk))
     }
 
     /** メッセージ一覧取得 */
     @GetMapping("/{talkId}/message")
     fun getIndividualTalk(
+        @AuthenticationPrincipal loginUser: Login,
         @PathVariable("talkId") talkId: String,
         messageId: String? = null
     ): MessagesResponse {
+        if(!talkService.hasTalk(talkId, loginUser.user.userId)) throw BadRequestException()
         val messages = messageService.getMessages(talkId, messageId)
         return MessagesResponse(messages?.map { MessageInfo(it) })
     }
@@ -49,11 +55,12 @@ class TalkAPIController(
     /** 個人トーク追加 */
     @PostMapping
     fun addIndividualTalk(
+        @AuthenticationPrincipal loginUser: Login,
         @Validated @RequestBody req: AddIndividualTalkRequest,
         bindingResult: BindingResult
     ): TalkResponse {
         if (bindingResult.hasErrors()) throw BadRequestException()
-        val insTalk = IndividualTalkIns(req.talkName, listOf(MockTestData.user1.userId, req.userId))
+        val insTalk = IndividualTalkIns(req.talkName, listOf(loginUser.user.userId, req.userId))
 
         val talk = talkService.addIndividualTalk(insTalk)
         return TalkResponse(TalkInfo(talk))
@@ -62,10 +69,12 @@ class TalkAPIController(
     /** トーク更新 */
     @PutMapping("/{talkId}")
     fun updTalk(
+        @AuthenticationPrincipal loginUser: Login,
         @PathVariable("talkId") talkId: String,
         @Validated @RequestBody req: UpdTalkRequest,
         bindingResult: BindingResult
     ): TalkResponse {
+        if(!talkService.hasTalk(talkId, loginUser.user.userId)) throw BadRequestException()
         if (bindingResult.hasErrors()) throw BadRequestException()
         val updTalk = TalkUpd(talkId, req.talkName)
 
@@ -76,20 +85,24 @@ class TalkAPIController(
     /** トーク削除 */
     @DeleteMapping("/{talkId}")
     fun deleteTalk(
+        @AuthenticationPrincipal loginUser: Login,
         @PathVariable("talkId") talkId: String
     ) {
+        if(!talkService.hasTalk(talkId, loginUser.user.userId)) throw BadRequestException()
         talkService.deleteTalk(talkId)
     }
 
     /** メッセージ送信 */
     @PostMapping("/{talkId}/message")
     fun postMessage(
+        @AuthenticationPrincipal loginUser: Login,
         @PathVariable("talkId") talkId: String,
         @Validated @RequestBody req: PostMessageRequest,
         bindingResult: BindingResult
     ): MessageResponse {
+        if(!talkService.hasTalk(talkId, loginUser.user.userId)) throw BadRequestException()
         if (bindingResult.hasErrors()) throw BadRequestException()
-        val insMessage = MessageIns(req.message, talkId, MockTestData.user1.userId)
+        val insMessage = MessageIns(req.message, talkId, loginUser.user.userId)
 
         val message = messageService.postMessage(insMessage)
         return MessageResponse(MessageInfo(message))
